requirements.txt
Django==4.2.7
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
Pillow==10.0.1
